Remove duplicate AddressFormProviders
