+++ test/controllers/register/InsurerAddressControllerSpec.scala
@@ -19,7 +19,7 @@ package controllers.register
 import connectors.FakeDataCacheConnector
 import controllers.ControllerSpecBase
 import controllers.actions._
-import forms.register.establishers.individual.AddressFormProvider
+import forms.address.AddressFormProvider
 import identifiers.register._
 import models.addresslookup.Address
 import models.register._
