+++ test/controllers/register/establishers/individual/AddressControllerSpec.scala
@@ -19,7 +19,7 @@ package controllers.register.establishers.individual
 import connectors.FakeDataCacheConnector
 import controllers.ControllerSpecBase
 import controllers.actions._
-import forms.register.establishers.individual.AddressFormProvider
+import forms.address.AddressFormProvider
 import identifiers.register.SchemeDetailsId
 import identifiers.register.establishers.individual.{AddressId, EstablisherDetailsId}
 import models.addresslookup.Address
