+++ test/views/register/InsurerAddressViewSpec.scala
@@ -17,7 +17,7 @@
 package views.register
 
 import controllers.register.routes
-import forms.register.establishers.individual.AddressFormProvider
+import forms.address.AddressFormProvider
 import models.NormalMode
 import models.addresslookup.Address
 import org.jsoup.Jsoup
