+++ test/views/register/establishers/company/CompanyPreviousAddressViewSpec.scala
@@ -16,9 +16,8 @@
 
 package views.register.establishers.company
 
-import forms.register.establishers.individual.AddressFormProvider
+import forms.address.AddressFormProvider
 import models.addresslookup.Address
-import models.register.CountryOptions
 import models.{Index, NormalMode}
 import play.api.data.Form
 import utils.InputOption
